refact or: Refactor API structure and implement chat and research modules
- Restructured main FastAPI application to include chat and research routers.
- Removed direct API key handling from main.py; centralized in respective services.
- Created chat module with models, service, and router for handling chat interactions.
- Developed research module with models, service, and router for research tasks.
- Updated Manim service to support asynchronous operations and improved code organization.
- Enhanced error handling and logging throughout the application.
- Streamlined AI interaction methods for both chat and research functionalities.
