+++ api/src/chat/__init__.py
@@ -1,0 +1,9 @@
+"""
+Chat module for handling conversations.
+"""
+
+from .models import ChatMessage, ChatRequest
+from .service import ChatService
+from .router import router
+
+__all__ = ["ChatMessage", "ChatRequest", "ChatService", "router"]
