+++ api/src/research/__init__.py
@@ -1,0 +1,9 @@
+"""
+Research module for handling research tasks.
+"""
+
+from .models import ResearchRequest, ResearchResponse, MultiAgentResearchResponse
+from .service import ResearchService
+from .router import router
+
+__all__ = ["ResearchRequest", "ResearchResponse", "MultiAgentResearchResponse", "ResearchService", "router"]
